Move graphql error handling logic to a separate class
